File reorganization, bunch of files moved to storage directory
