+++ companion/src/CMakeLists.txt
@@ -75,12 +75,14 @@ include_directories(
   ${COMPANION_SRC_DIRECTORY}/thirdparty/qxtcommandoptions
   ${COMPANION_SRC_DIRECTORY}/thirdparty/qcustomplot
   ${COMPANION_SRC_DIRECTORY}/thirdparty/miniz
+  ${COMPANION_SRC_DIRECTORY}/storage
 )
 
 add_subdirectory(shared)
 add_subdirectory(modeledit)
 add_subdirectory(generaledit)
 add_subdirectory(simulation)
+add_subdirectory(storage)
 add_subdirectory(thirdparty/qcustomplot)
 add_subdirectory(thirdparty/qxtcommandoptions)
 
@@ -99,8 +101,6 @@ set(common_SRCS
   firmwares/opentx/open9xGruvin9xeeprom.cpp
   firmwares/opentx/open9xSky9xeeprom.cpp
   firmwares/opentx/opentxinterface.cpp
-  file.cpp
-  appdata.cpp
   )
 
 qt5_wrap_cpp(common_SRCS)
@@ -109,16 +109,10 @@ add_library(common ${common_SRCS})
 qt5_use_modules(common Core Xml Widgets)
 
 set(companion_SRCS
-  hexinterface.cpp
-  firmwareinterface.cpp
-  converteeprom.cpp
-  # xmlinterface.cpp
-  # ${PROJECT_BINARY_DIR}/radio.cxx
   helpers.cpp
   helpers_html.cpp
   mdichild.cpp
   modelslist.cpp
-  mountlist.cpp
   apppreferencesdialog.cpp
   fwpreferencesdialog.cpp
   burnconfigdialog.cpp
@@ -238,7 +232,7 @@ add_executable(${COMPANION_NAME} MACOSX_BUNDLE WIN32 ${companion_SRCS} ${compani
 
 add_dependencies(${COMPANION_NAME} gen_qrc)
 qt5_use_modules(${COMPANION_NAME} Core Widgets Network)
-target_link_libraries(${COMPANION_NAME} PRIVATE generaledit modeledit simulation common qcustomplot shared ${PTHREAD_LIBRARY} ${SDL_LIBRARY})
+target_link_libraries(${COMPANION_NAME} PRIVATE generaledit modeledit simulation common qcustomplot shared storage ${PTHREAD_LIBRARY} ${SDL_LIBRARY})
 
 ############# Standalone simu ###############
 
@@ -263,7 +257,7 @@ add_executable(${SIMULATOR_NAME} MACOSX_BUNDLE WIN32 ${simu_SRCS} ${companion_RC
 
 add_dependencies(${SIMULATOR_NAME} gen_qrc)
 # TODO not the same link command than companion?
-target_link_libraries(${SIMULATOR_NAME} PRIVATE simulation common qxtcommandoptions ${QT_LIBRARIES} ${QT_QTMAIN_LIBRARY} ${PTHREAD_LIBRARY} ${SDL_LIBRARY} ${OPENTX_SIMULATOR_LIBS})
+target_link_libraries(${SIMULATOR_NAME} PRIVATE simulation common storage qxtcommandoptions ${QT_LIBRARIES} ${QT_QTMAIN_LIBRARY} ${PTHREAD_LIBRARY} ${SDL_LIBRARY} ${OPENTX_SIMULATOR_LIBS})
 
 ############# Packaging ####################
 message(STATUS ${CMAKE_BINARY_DIR} ${CMAKE_INSTALL_PREFIX}/bin)
