--- companion/src/converteeprom.h
@@ -1,8 +1,0 @@
-#ifndef CONVERTHEADER_H_
-#define CONVERTHEADER_H_
-
-#include <QString>
-
-bool convertEEprom(const QString &sourceEEprom, const QString &destinationEEprom, const QString &firmware);
-
-#endif /* CONVERTHEADER_H_ */
