+++ companion/src/firmwares/gruvin9x/gruvin9xinterface.cpp
@@ -17,7 +17,7 @@
 #include <iostream>
 #include "gruvin9xinterface.h"
 #include "gruvin9xeeprom.h"
-#include "file.h"
+#include "rlefile.h"
 #include "appdata.h"
 
 #define FILE_TYP_GENERAL 1
