+++ companion/src/firmwares/th9x/th9xinterface.cpp
@@ -17,7 +17,7 @@
 #include <iostream>
 #include "th9xinterface.h"
 #include "th9xeeprom.h"
-#include "file.h"
+#include "rlefile.h"
 #include "appdata.h"
 
 #define FILE_TYP_GENERAL 1
