+++ companion/src/storage/CMakeLists.txt
@@ -1,0 +1,26 @@
+include_directories(${CMAKE_SOURCE_DIR} ${CMAKE_CURRENT_BINARY_DIR})
+
+set(storage_NAMES
+  storage_eeprom
+  storage_sdcard
+  hexinterface
+  # xmlinterface
+  mountlist
+  rlefile
+  appdata
+  firmwareinterface
+)
+
+set(storage_SRCS
+)
+
+set(storage_HDRS
+)
+
+foreach(name ${storage_NAMES})
+  set(storage_SRCS ${storage_SRCS} ${name}.cpp)
+  set(storage_HDRS ${storage_HDRS} ${name}.h)
+endforeach()
+
+add_library(storage ${storage_SRCS})
+target_link_libraries(storage Qt5::Core Qt5::Gui Qt5::Xml Qt5::Widgets)
