+++ companion/src/storage/appdata.h
