+++ companion/src/storage/firmwareinterface.cpp
