+++ companion/src/storage/hexinterface.cpp
