+++ companion/src/storage/hexinterface.h
