+++ companion/src/storage/storage_sdcard.h
@@ -1,0 +1,25 @@
+/*
+ * Author - Bertrand Songis <bsongis@gmail.com>
+ *
+ * Based on th9x -> http://code.google.com/p/th9x/
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 2 as
+ * published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ */
+
+#ifndef storage_sdcard_h
+#define storage_sdcard_h
+
+class StorageSdcard {
+  pubilic:
+  private:
+};
+
+#endif
