+++ companion/src/storage/xmlinterface.cpp
