+++ companion/src/storage/xmlinterface.h
